iteration 9
refactored code, added separated ProgressScreen in to ProgressView and ProgressScreen, added ProgressScreens for Calories, Macro and Weight and the corelating view and context files for each. Started trying to work on the weight progress graph but need to do further research. Added ProfileScreen and ProfileView aka. the Profile settings page, where users can change height, age, weight and sex, and from a dropdown they can chose to use either metric or imperial as their measurement system. Still have yet to iron out some bugs with this, but that will be done in the next iteration.